+++ sae-20190506/alibabacloud_sae20190506/models/_exec_job_request.py
@@ -18,11 +18,41 @@ class ExecJobRequest(DaraModel):
         time: str = None,
         war_start_options: str = None,
     ):
+        # The job template ID.
+        # 
         # This parameter is required.
         self.app_id = app_id
+        # The command that is used to start the image. The command must be an existing executable object in the container. Example:
+        # 
+        #     command:
+        #           - echo
+        #           - abc
+        #           - >
+        #           - file0
+        # 
+        # In this example, the Command parameter is set to `Command="echo", CommandArgs=["abc", ">", "file0"]`.
         self.command = command
+        # The parameters of the image startup command. The CommandArgs parameter specifies the parameters that are required for the **Command** parameter. The name must meet the following format requirements:
+        # 
+        # `["a","b"]`
+        # 
+        # In the preceding example, the CommandArgs parameter is set to `CommandArgs=["abc", ">", "file0"]`. The data type of `["abc", ">", "file0"]` must be an array of strings in the JSON format. This parameter is optional.
         self.command_args = command_args
+        # The environment variables. You can configure custom environment variables or reference a ConfigMap. If you want to reference a ConfigMap, you must first create a ConfigMap. For more information, see [CreateConfigMap](https://help.aliyun.com/document_detail/176914.html). Valid values:
+        # 
+        # *   Configure custom environment variables
+        # 
+        #     *   **name**: the name of the environment variable.
+        #     *   **value**: the value of the environment variable.
+        # 
+        # *   Reference ConfigMap
+        # 
+        #     *   **name**: the name of the environment variable. You can reference one or all keys. If you want to reference all keys, specify `sae-sys-configmap-all-<ConfigMap name>`. Example: `sae-sys-configmap-all-test1`.
+        #     *   **valueFrom**: the reference of the environment variable. Set the value to `configMapRef`.
+        #     *   **configMapId**: the ConfigMap ID.
+        #     *   **key**: the key. If you want to reference all keys, do not configure this parameter.
         self.envs = envs
+        # The event ID. This is a user-defined parameter used for idempotency so that only one job is created for the same event ID.
         self.event_id = event_id
         # The arguments in the JAR package. The arguments are used to start the job. The default startup command is `$JAVA_HOME/bin/java $JarStartOptions -jar $CATALINA_OPTS "$package_path" $JarStartArgs`.
         self.jar_start_args = jar_start_args