+++ sae-20190506/alibabacloud_sae20190506/models/_exec_job_response_body.py
@@ -16,12 +16,33 @@ class ExecJobResponseBody(DaraModel):
         success: bool = None,
         trace_id: str = None,
     ):
+        # The HTTP status code. Valid values: 
+        # 
+        # *   **2xx**: The call was successful.
+        # *   **3xx**: The call was redirected.
+        # *   **4xx**: The call failed.
+        # *   **5xx**: A server error occurred.
         self.code = code
+        # The returned data.
         self.data = data
+        # The error code. Valid values:
+        # 
+        # *   If the call is successful, the **ErrorCode** parameter is not returned.
+        # *   If the call fails, the **ErrorCode** parameter is returned. For more information, see the **Error codes** section in this topic.
         self.error_code = error_code
+        # The returned message. Valid values:
+        # 
+        # *   success: If the call is successful, **success** is returned.
+        # *   An error code: If the call fails, an error code is returned.
         self.message = message
+        # The request ID.
         self.request_id = request_id
+        # Indicates whether the request was successful. Valid values:
+        # 
+        # *   **true**
+        # *   **false**
         self.success = success
+        # The ID of the trace. This parameter is used to query the exact call information.
         self.trace_id = trace_id
 
     def validate(self):
@@ -90,9 +111,24 @@ class ExecJobResponseBodyData(DaraModel):
         msg: str = None,
         success: str = None,
     ):
+        # The HTTP status code. Valid values: 
+        # 
+        # *   **2xx**: The call was successful.
+        # *   **3xx**: The call was redirected.
+        # *   **4xx**: The call failed.
+        # *   **5xx**: A server error occurred.
         self.code = code
+        # The job ID.
         self.data = data
+        # The returned message. Valid values:
+        # 
+        # - success: If the call is successful, **success** is returned.
+        # - An error code: If the call fails, an error code is returned.
         self.msg = msg
+        # Indicates whether the request was successful. Valid values:
+        # 
+        # *   **true**
+        # *   **false**
         self.success = success
 
     def validate(self):